backend_django/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
dj-database-url==2.1.0
python-dotenv==1.0.1